feat: add EthFoxStaking provider
