+++ src/features/defi/contexts/DefiManagerProvider/DefiCommon.ts
@@ -11,6 +11,7 @@ export enum DefiProvider {
   Idle = 'idle',
   Yearn = 'yearn',
   ShapeShift = 'ShapeShift',
+  EthFoxStaking = 'ETH/FOX Staking',
   UniV2 = 'Uniswap V2',
   Cosmos = 'Cosmos',
   Osmosis = 'Osmosis',
